demoreal/requirements.txt
pillow>=10.0.0
RPi.GPIO>=0.7.0; platform_system=="Linux"